mcp_server/requirements.txt
mcp>=1.0.0
pymongo>=4.6.0
python-dotenv>=1.0.0
pydantic>=2.5.0
asyncio>=3.4.3
bson
typing
